08/09/2025
Experimented with a 'obj_party_data' to store party data in an object for easier manipulation; didn't turn out so well so I reverted to the previous commit.
Also made it so that AP is full at the beginning of each fight (aka made the save system NOT save/load the AP variable).