25/09/2025
Items and Abilities now get greyed out if they are unable to be used in battle (e.g. item quantity is 0, or not enough AP to use ability).
Item array system has been changed to be initialised with all possible items (amount set to 0) from the start; this makes it easier to use without having to use for loops to find the specific location of an item in the array.
TO-DO:
- Fix item inventory (make items invisible if the player doesn't have any, make sure the slots are assigned correctly, etc.)
- Check over item code in other objects to make sure it follows the new system's standards