10/10/2025
Updated 'battle_change_ap' function to optionally take the target's Agility stat then use that in the formula '_amount = floor((_target.agility + 2) / 3);' to determine how much AP to restore. This allows for a raw amount to be manually entered (e.g. when deducting the AP cost of using an ability), or for the amount to be automatically calculated based on the target's Agility stat (e.g. when restoring AP at the beginning of the target's combat turn).

Replaced the old turn order system (shuffling the 'units' array randomly) with a sorting system that puts units in order of their Agility stat (faster units go sooner).