16/10/2025
Added entity collisions.
Fine-tuned tile collision to account for the player's sprite.
Added y-based depth to entities and player (if the player moves behind an entity, the player will be drawn behind the entity and vice versa).
Added enemies as children to the entity parent object.
Fixed keycard item not being able to be picked up.