04/11/2025
Added additional text box functionality and updated pre-existing function calls to utilise it: The size and position of a text box can now be defined in the function call; this allows for different events to create different types of text boxes (e.g. talking with someone can create a large 'Dialogue' text box to fit responses, and using an item in the inventory can create a smaller 'Inventory' text box that doesn't take up as much space).