13/11/2025
Planned future updates, added them to Notes file.
Limited levelling up SPECIAL stats to a maximum of 10.
Balanced damage formulas.
Started work on a new 'Swap' screen which will allow the player to swap around which party members are in the active party (for combat).
Selected a party member in the inventory screen now prompts the user to select whether they want to see that member's 'Stats' or if they want to 'Swap' them out.